Update test projects, enum assignments, and add edge tests
- Use tabs for indentation in .csproj files and add <Exclude>GenDI</Exclude> to test projects
- Remove Microsoft.Testing.Extensions.CodeCoverage from some test projects
- Add <ExcludeFromCodeCoverage>true</ExcludeFromCodeCoverage> to GenDI.csproj
- Assign ThreadIsolationPolicy enum values directly from ServiceLifetime
- Add Roslyn fallback and pragma warning disable in AnalysisPrivateMethodsTests
- Add tests for option config without consumers and decorator without implementation in GeneratorEdgeCaseTests

+++ src/GenDI/ThreadIsolationPolicy.cs
@@ -15,15 +15,15 @@ public enum ThreadIsolationPolicy
     /// <summary>
     /// Enables thread-local caching with singleton registration semantics.
     /// </summary>
-    Singleton = (int)ServiceLifetime.Singleton,
+    Singleton = ServiceLifetime.Singleton,
 
     /// <summary>
     /// Enables thread-local caching with scoped registration semantics.
     /// </summary>
-    Scoped = (int)ServiceLifetime.Scoped,
+    Scoped = ServiceLifetime.Scoped,
 
     /// <summary>
     /// Enables thread-local caching with transient registration semantics.
     /// </summary>
-    Transient = (int)ServiceLifetime.Transient,
+    Transient = ServiceLifetime.Transient,
 }
